update: update get gas in metis transaction

+++ lib/core/ethereum/ethereum_transaction_simulate_result.dart
@@ -2,7 +2,7 @@ import 'package:web3dart/web3dart.dart';
 
 class EthereumTransactionSimulateResult{
   Transaction transaction;
-  int? gas;
+  BigInt? gas;
 
   EthereumTransactionSimulateResult({
     required this.transaction,
